% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askUserYesNo.R
\name{askUserYesNo}
\alias{askUserYesNo}
\title{Ask user for a yes/no response}
\usage{
askUserYesNo(prompt, interactive.only = TRUE)
}
\arguments{
\item{prompt}{\code{character()} A close-ended yes or no question to pose
to the user e.g., \code{"Do you want to continue?"}. Note that the question
must include a question mark (\verb{?}).}

\item{interactive.only}{\code{logical(1)} If \code{TRUE}, the function will only
prompt the user when the R session is interactive. If \code{FALSE}, the
function will always prompt the user.}
}
\value{
TRUE when user replies with 'yes' to prompt, FALSE when 'no'
}
\description{
Ask user for a yes/no response
}
\examples{

askUserYesNo("Do you want to continue?")

}
\author{
Martin M.
}
