% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocBaseUtils-package.R
\docType{package}
\name{BiocBaseUtils-package}
\alias{BiocBaseUtils}
\alias{BiocBaseUtils-package}
\title{BiocBaseUtils: Utility and internal functions for Bioconductor packages}
\description{
The package coalesces typical helper functions that are scattered throughout the Bioconductor ecosystem. It aims to reduce code redundancy by formalizing functions often used by Bioconductor developers. These functions include operations such as replacing slots in an object, selecting observations for show methods, labeling function life cycles, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/Bioconductor/BiocBaseUtils}
  \item Report bugs at \url{https://www.github.com/Bioconductor/BiocBaseUtils/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})
}

Other contributors:
\itemize{
  \item Martin Morgan \email{martin.morgan@roswellpark.org} [contributor]
  \item Hervé Pagès \email{hpages.on.github@gmail.com} [contributor]
}

}
\keyword{internal}
