% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocUtils-package.R
\docType{package}
\name{BiocUtils-package}
\title{BiocUtils: Utility functions for developing Bioconductor packages}
\description{
BiocUtils attempts to formalize often written developer functions that
are scattered throughout the Bioconductor ecosystem. Some of these functions
include the ability to replace slots in an object. Other functions work
to create a nice show method output by selecting some observations. This
package also allows developers to check what emails are associated with
what packages.
}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@roswellpark.org} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Other contributors:
\itemize{
  \item Hervé Pagès \email{hpages.on.github@gmail.com} [contributor]
}

}
