% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocDeveloper-package.R
\docType{package}
\name{BiocBaseUtils-package}
\alias{BiocBaseUtils}
\alias{BiocBaseUtils-package}
\title{BiocBaseUtils: Utility and Internal functions for Bioconductor packages}
\description{
BiocBaseUtils is a package aimed at helping the typical Bioconductor developer
formalize often written functions that can be seen scattered throughout the
Bioconductor ecosystem. Some of these functions include the ability to
replace slots in an object. Other functions work to create a nice show method
output by selecting some observations.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://www.github.com/Bioconductor/BiocBaseUtils/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@roswellpark.org} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Other contributors:
\itemize{
  \item Martin Morgan \email{martin.morgan@roswellpark.org} [contributor]
  \item Hervé Pagès \email{hpages.on.github@gmail.com} [contributor]
}

}
